+++ src/plugin/bf.ts
@@ -277,8 +277,6 @@ const help_text = `<code>${mainPrefix}bf</code> 备份 plugins + assets 目录
 
 // 插件类
 class BfPlugin extends Plugin {
-  ignoreEdited: boolean = true;
-
   description = `\n📦 备份插件\n\n${help_text}
 
 若想实现定时备份, 可安装并使用 <code>${mainPrefix}tpm i acron</code>