+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.2.5] --2025-09-21
 
 - Plugin 添加 ignoreEdited 参数，默认为 `true`, 即插件不监听编辑信息事件
+  - 命令监听忽略编辑的消息: `true` (可使用环境变量 `TB_CMD_IGNORE_EDITED` 覆盖)
 - eatgif 利用 ffmpeg 将 gif 转成 webm 表情包
 - 添加 dotenv 依赖 以及 env 配置文件